v2extension/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
pgvector>=0.2.4
redis[hiredis]>=5.0.1
aioredis>=2.0.1
aio-pika>=9.3.1
langchain>=0.1.0
langchain-openai>=0.0.2
langgraph>=0.0.20
langsmith>=0.0.77
openai>=1.10.0
minio>=7.2.3
boto3>=1.34.29
pyarrow>=15.0.0
pandas>=2.2.0
websockets>=12.0
python-socketio>=5.11.0
python-dotenv>=1.0.0
httpx>=0.26.0
tenacity>=8.2.3
pytest>=8.0.0
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
